refactor: Move asProcessArg in utils
