+++ src/types/global.ts
@@ -20,13 +20,3 @@ export interface Processes {
     sensor: ChildProcess | null;
     stream: ChildProcess | null;
 }
-
-/**
- * Check if process argument is present
- *
- * @param {string} argument - Argument to check
- * @returns {boolean} - True if argument is present
- */
-export function asProcessArg(argument: string): boolean {
-    return process.argv.slice(2).includes(argument);
-}
